Draft: move Dimension GuiCommand to gui_dimensions module
