+++ src/Mod/Draft/DraftTools.py
@@ -162,360 +162,7 @@ from draftguitools.gui_circles import Circle
 from draftguitools.gui_polygons import Polygon
 from draftguitools.gui_ellipses import Ellipse
 from draftguitools.gui_texts import Text
-
-
-class Dimension(Creator):
-    """The Draft_Dimension FreeCAD command definition"""
-
-    def __init__(self):
-        self.max=2
-        self.cont = None
-        self.dir = None
-
-    def GetResources(self):
-        return {'Pixmap'  : 'Draft_Dimension',
-                'Accel' : "D, I",
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Dimension", "Dimension"),
-                'ToolTip': QtCore.QT_TRANSLATE_NOOP("Draft_Dimension", "Creates a dimension. CTRL to snap, SHIFT to constrain, ALT to select a segment")}
-
-    def Activated(self):
-        name = translate("draft","Dimension")
-        if self.cont:
-            self.finish()
-        elif self.hasMeasures():
-            Creator.Activated(self,name)
-            self.dimtrack = trackers.dimTracker()
-            self.arctrack = trackers.arcTracker()
-            self.createOnMeasures()
-            self.finish()
-        else:
-            Creator.Activated(self,name)
-            if self.ui:
-                self.ui.pointUi(name)
-                self.ui.continueCmd.show()
-                self.ui.selectButton.show()
-                self.altdown = False
-                self.call = self.view.addEventCallback("SoEvent",self.action)
-                self.dimtrack = trackers.dimTracker()
-                self.arctrack = trackers.arcTracker()
-                self.link = None
-                self.edges = []
-                self.pts = []
-                self.angledata = None
-                self.indices = []
-                self.center = None
-                self.arcmode = False
-                self.point2 = None
-                self.force = None
-                self.info = None
-                self.selectmode = False
-                self.setFromSelection()
-                FreeCAD.Console.PrintMessage(translate("draft", "Pick first point")+"\n")
-                FreeCADGui.draftToolBar.show()
-
-    def setFromSelection(self):
-        """If we already have selected geometry, fill the nodes accordingly"""
-        sel = FreeCADGui.Selection.getSelectionEx()
-        import DraftGeomUtils
-        if len(sel) == 1:
-            if len(sel[0].SubElementNames) == 1:
-                if "Edge" in sel[0].SubElementNames[0]:
-                    edge = sel[0].SubObjects[0]
-                    n = int(sel[0].SubElementNames[0].lstrip("Edge"))-1
-                    self.indices.append(n)
-                    if DraftGeomUtils.geomType(edge) == "Line":
-                        self.node.extend([edge.Vertexes[0].Point,edge.Vertexes[1].Point])
-                        v1 = None
-                        v2 =None
-                        for i,v in enumerate(sel[0].Object.Shape.Vertexes):
-                            if v.Point == edge.Vertexes[0].Point:
-                                v1 = i
-                            if v.Point == edge.Vertexes[1].Point:
-                                v2 = i
-                        if (v1 != None) and (v2 != None):
-                            self.link = [sel[0].Object,v1,v2]
-                    elif DraftGeomUtils.geomType(edge) == "Circle":
-                        self.node.extend([edge.Curve.Center,edge.Vertexes[0].Point])
-                        self.edges = [edge]
-                        self.arcmode = "diameter"
-                        self.link = [sel[0].Object,n]
-
-    def hasMeasures(self):
-        """checks if only measurements objects are selected"""
-        sel = FreeCADGui.Selection.getSelection()
-        if not sel:
-            return False
-        for o in sel:
-            if not o.isDerivedFrom("App::MeasureDistance"):
-                return False
-        return True
-
-    def finish(self,closed=False):
-        """terminates the operation"""
-        self.cont = None
-        self.dir = None
-        Creator.finish(self)
-        if self.ui:
-            self.dimtrack.finalize()
-            self.arctrack.finalize()
-
-    def createOnMeasures(self):
-        for o in FreeCADGui.Selection.getSelection():
-            p1 = o.P1
-            p2 = o.P2
-            pt = o.ViewObject.RootNode.getChildren()[1].getChildren()[0].getChildren()[0].getChildren()[3]
-            p3 = Vector(pt.point.getValues()[2].getValue())
-            FreeCADGui.addModule("Draft")
-            self.commit(translate("draft","Create Dimension"),
-                        ['dim = Draft.makeDimension('+DraftVecUtils.toString(p1)+','+DraftVecUtils.toString(p2)+','+DraftVecUtils.toString(p3)+')',
-                         'FreeCAD.ActiveDocument.removeObject("'+o.Name+'")',
-                         'Draft.autogroup(dim)',
-                         'FreeCAD.ActiveDocument.recompute()'])
-
-    def createObject(self):
-        """creates an object in the current doc"""
-        FreeCADGui.addModule("Draft")
-        if self.angledata:
-            normal = "None"
-            if len(self.edges) == 2:
-                import DraftGeomUtils
-                v1 = DraftGeomUtils.vec(self.edges[0])
-                v2 = DraftGeomUtils.vec(self.edges[1])
-                normal = DraftVecUtils.toString((v1.cross(v2)).normalize())
-            self.commit(translate("draft","Create Dimension"),
-                        ['dim = Draft.makeAngularDimension(center='+DraftVecUtils.toString(self.center)+',angles=['+str(self.angledata[0])+','+str(self.angledata[1])+'],p3='+DraftVecUtils.toString(self.node[-1])+',normal='+normal+')',
-                        'Draft.autogroup(dim)',
-                        'FreeCAD.ActiveDocument.recompute()'])
-        elif self.link and (not self.arcmode):
-            ops = []
-            if self.force == 1:
-                self.commit(translate("draft","Create Dimension"),
-                        ['dim = Draft.makeDimension(FreeCAD.ActiveDocument.'+self.link[0].Name+','+str(self.link[1])+','+str(self.link[2])+','+DraftVecUtils.toString(self.node[2])+')','dim.Direction=FreeCAD.Vector(0,1,0)',
-                        'Draft.autogroup(dim)',
-                        'FreeCAD.ActiveDocument.recompute()'])
-            elif self.force == 2:
-                self.commit(translate("draft","Create Dimension"),
-                        ['dim = Draft.makeDimension(FreeCAD.ActiveDocument.'+self.link[0].Name+','+str(self.link[1])+','+str(self.link[2])+','+DraftVecUtils.toString(self.node[2])+')','dim.Direction=FreeCAD.Vector(1,0,0)',
-                        'Draft.autogroup(dim)',
-                        'FreeCAD.ActiveDocument.recompute()'])
-            else:
-                self.commit(translate("draft","Create Dimension"),
-                        ['dim = Draft.makeDimension(FreeCAD.ActiveDocument.'+self.link[0].Name+','+str(self.link[1])+','+str(self.link[2])+','+DraftVecUtils.toString(self.node[2])+')',
-                        'Draft.autogroup(dim)',
-                        'FreeCAD.ActiveDocument.recompute()'])
-        elif self.arcmode:
-            self.commit(translate("draft","Create Dimension"),
-                        ['dim = Draft.makeDimension(FreeCAD.ActiveDocument.'+self.link[0].Name+','+str(self.link[1])+',"'+str(self.arcmode)+'",'+DraftVecUtils.toString(self.node[2])+')',
-                        'Draft.autogroup(dim)',
-                        'FreeCAD.ActiveDocument.recompute()'])
-        else:
-            self.commit(translate("draft","Create Dimension"),
-                        ['dim = Draft.makeDimension('+DraftVecUtils.toString(self.node[0])+','+DraftVecUtils.toString(self.node[1])+','+DraftVecUtils.toString(self.node[2])+')',
-                        'Draft.autogroup(dim)',
-                        'FreeCAD.ActiveDocument.recompute()'])
-        if self.ui.continueMode:
-            self.cont = self.node[2]
-            if not self.dir:
-                if self.link:
-                    v1 = self.link[0].Shape.Vertexes[self.link[1]].Point
-                    v2 = self.link[0].Shape.Vertexes[self.link[2]].Point
-                    self.dir = v2.sub(v1)
-                else:
-                    self.dir = self.node[1].sub(self.node[0])
-            self.node = [self.node[1]]
-        self.link = None
-
-    def selectEdge(self):
-        self.selectmode = not(self.selectmode)
-
-    def action(self,arg):
-        """scene event handler"""
-        if arg["Type"] == "SoKeyboardEvent":
-            if arg["Key"] == "ESCAPE":
-                self.finish()
-        elif arg["Type"] == "SoLocation2Event": #mouse movement detection
-            import DraftGeomUtils
-            shift = hasMod(arg,MODCONSTRAIN)
-            if self.arcmode or self.point2:
-                setMod(arg,MODCONSTRAIN,False)
-            self.point,ctrlPoint,self.info = getPoint(self,arg,noTracker=(len(self.node)>0))
-            if (hasMod(arg,MODALT) or self.selectmode) and (len(self.node)<3):
-                self.dimtrack.off()
-                if not self.altdown:
-                    self.altdown = True
-                    self.ui.switchUi(True)
-                    if hasattr(FreeCADGui,"Snapper"):
-                        FreeCADGui.Snapper.setSelectMode(True)
-                snapped = self.view.getObjectInfo((arg["Position"][0],arg["Position"][1]))
-                if snapped:
-                    ob = self.doc.getObject(snapped['Object'])
-                    if "Edge" in snapped['Component']:
-                        num = int(snapped['Component'].lstrip('Edge'))-1
-                        ed = ob.Shape.Edges[num]
-                        v1 = ed.Vertexes[0].Point
-                        v2 = ed.Vertexes[-1].Point
-                        self.dimtrack.update([v1,v2,self.cont])
-            else:
-                if self.node and (len(self.edges) < 2):
-                    self.dimtrack.on()
-                if len(self.edges) == 2:
-                    # angular dimension
-                    self.dimtrack.off()
-                    r = self.point.sub(self.center)
-                    self.arctrack.setRadius(r.Length)
-                    a = self.arctrack.getAngle(self.point)
-                    pair = DraftGeomUtils.getBoundaryAngles(a,self.pts)
-                    if not (pair[0] < a < pair[1]):
-                        self.angledata = [4*math.pi-pair[0],2*math.pi-pair[1]]
-                    else:
-                        self.angledata = [2*math.pi-pair[0],2*math.pi-pair[1]]
-                    self.arctrack.setStartAngle(self.angledata[0])
-                    self.arctrack.setEndAngle(self.angledata[1])
-                if self.altdown:
-                    self.altdown = False
-                    self.ui.switchUi(False)
-                    if hasattr(FreeCADGui,"Snapper"):
-                        FreeCADGui.Snapper.setSelectMode(False)
-                if self.dir:
-                    self.point = self.node[0].add(DraftVecUtils.project(self.point.sub(self.node[0]),self.dir))
-                if len(self.node) == 2:
-                    if self.arcmode and self.edges:
-                        cen = self.edges[0].Curve.Center
-                        rad = self.edges[0].Curve.Radius
-                        baseray = self.point.sub(cen)
-                        v2 = DraftVecUtils.scaleTo(baseray,rad)
-                        v1 = v2.negative()
-                        if shift:
-                            self.node = [cen,cen.add(v2)]
-                            self.arcmode = "radius"
-                        else:
-                            self.node = [cen.add(v1),cen.add(v2)]
-                            self.arcmode = "diameter"
-                        self.dimtrack.update(self.node)
-                # Draw constraint tracker line.
-                if shift and (not self.arcmode):
-                    if len(self.node) == 2:
-                        if not self.point2:
-                            self.point2 = self.node[1]
-                        else:
-                            self.node[1] = self.point2
-                        if not self.force:
-                            a=abs(self.point.sub(self.node[0]).getAngle(plane.u))
-                            if (a > math.pi/4) and (a <= 0.75*math.pi):
-                                self.force = 1
-                            else:
-                                self.force = 2
-                        if self.force == 1:
-                            self.node[1] = Vector(self.node[0].x,self.node[1].y,self.node[0].z)
-                        elif self.force == 2:
-                            self.node[1] = Vector(self.node[1].x,self.node[0].y,self.node[0].z)
-                else:
-                    self.force = None
-                    if self.point2 and (len(self.node) > 1):
-                        self.node[1] = self.point2
-                        self.point2 = None
-                # update the dimline
-                if self.node and (not self.arcmode):
-                    self.dimtrack.update(self.node+[self.point]+[self.cont])
-            redraw3DView()
-        elif arg["Type"] == "SoMouseButtonEvent":
-            if (arg["State"] == "DOWN") and (arg["Button"] == "BUTTON1"):
-                import DraftGeomUtils
-                if self.point:
-                    self.ui.redraw()
-                    if (not self.node) and (not self.support):
-                        getSupport(arg)
-                    if (hasMod(arg,MODALT) or self.selectmode) and (len(self.node)<3):
-                        #print("snapped: ",self.info)
-                        if self.info:
-                            ob = self.doc.getObject(self.info['Object'])
-                            if 'Edge' in self.info['Component']:
-                                num = int(self.info['Component'].lstrip('Edge'))-1
-                                ed = ob.Shape.Edges[num]
-                                v1 = ed.Vertexes[0].Point
-                                v2 = ed.Vertexes[-1].Point
-                                i1 = i2 = None
-                                for i in range(len(ob.Shape.Vertexes)):
-                                    if v1 == ob.Shape.Vertexes[i].Point:
-                                        i1 = i
-                                    if v2 == ob.Shape.Vertexes[i].Point:
-                                        i2 = i
-                                if (i1 != None) and (i2 != None):
-                                    self.indices.append(num)
-                                    if not self.edges:
-                                        # nothing snapped yet, we treat it as normal edge-snapped dimension
-                                        self.node = [v1,v2]
-                                        self.link = [ob,i1,i2]
-                                        self.edges.append(ed)
-                                        if DraftGeomUtils.geomType(ed) == "Circle":
-                                            # snapped edge is an arc
-                                            self.arcmode = "diameter"
-                                            self.link = [ob,num]
-                                    else:
-                                        # there is already a snapped edge, so we start angular dimension
-                                        self.edges.append(ed)
-                                        self.node.extend([v1,v2]) # self.node now has the 4 endpoints
-                                        c = DraftGeomUtils.findIntersection(self.node[0],
-                                                                   self.node[1],
-                                                                   self.node[2],
-                                                                   self.node[3],
-                                                                   True,True)
-                                        if c:
-                                            #print("centers:",c)
-                                            self.center = c[0]
-                                            self.arctrack.setCenter(self.center)
-                                            self.arctrack.on()
-                                            for e in self.edges:
-                                                for v in e.Vertexes:
-                                                    self.pts.append(self.arctrack.getAngle(v.Point))
-                                            self.link = [self.link[0],ob]
-                                        else:
-                                            FreeCAD.Console.PrintMessage(translate("draft", "Edges don't intersect!")+"\n")
-                                            self.finish()
-                                            return
-                                self.dimtrack.on()
-                    else:
-                        self.node.append(self.point)
-                    self.selectmode = False
-                    #print("node",self.node)
-                    self.dimtrack.update(self.node)
-                    if (len(self.node) == 2):
-                        self.point2 = self.node[1]
-                    if (len(self.node) == 1):
-                        self.dimtrack.on()
-                        if self.planetrack:
-                            self.planetrack.set(self.node[0])
-                    elif (len(self.node) == 2) and self.cont:
-                        self.node.append(self.cont)
-                        self.createObject()
-                        if not self.cont:
-                            self.finish()
-                    elif (len(self.node) == 3):
-                        # for unlinked arc mode:
-                        # if self.arcmode:
-                        #        v = self.node[1].sub(self.node[0])
-                        #        v.multiply(0.5)
-                        #        cen = self.node[0].add(v)
-                        #        self.node = [self.node[0],self.node[1],cen]
-                        self.createObject()
-                        if not self.cont:
-                            self.finish()
-                    elif self.angledata:
-                        self.node.append(self.point)
-                        self.createObject()
-                        self.finish()
-
-    def numericInput(self,numx,numy,numz):
-        """this function gets called by the toolbar when valid x, y, and z have been entered there"""
-        self.point = Vector(numx,numy,numz)
-        self.node.append(self.point)
-        self.dimtrack.update(self.node)
-        if (len(self.node) == 1):
-            self.dimtrack.on()
-        elif (len(self.node) == 3):
-            self.createObject()
-            if not self.cont:
-                self.finish()
+from draftguitools.gui_dimensions import Dimension
 
 
 class ShapeString(Creator):
@@ -3353,7 +3000,6 @@ from draftguitools.gui_snaps import ShowSnapBar
 #---------------------------------------------------------------------------
 
 # drawing commands
-FreeCADGui.addCommand('Draft_Dimension',Dimension())
 FreeCADGui.addCommand('Draft_Point',Point())
 FreeCADGui.addCommand('Draft_ShapeString',ShapeString())
 FreeCADGui.addCommand('Draft_Facebinder',Draft_Facebinder())
